Holy genuine shit.
The robot wouldn't go backwards or to the left. The similarity between the two directions was that they both relied on a negative value for the joystick positioning (controller.axis_.position()). Joystick positioning is on a scale of -100 to 100, and I was using the function f(x) = 0.01x^2 + x to convert joystick position (x) to RPM for the motors. In this function, positive values went continually up, while negative values had a peak RPM of 25 when x = -50. After that, f(x) continues to approach 0 until x = -100, where f(x) is 0. This, in turn, slowed motor speeds as x decreased towards -100. I've been ripping my goddamn hair out for almost a week over this. Over a MATH error.

- Arymus Reyes, sophomore and algebra 2 student (ironically).